I can use commas, realy, I swear. I tidied up the validation for slots a little but it still doesn't work because valid_tutor isn't called when you add a tutor to a slot, apparently.
It probably has something to do with the fact that it's not that the actual slot
is getting changed, but a record in the join table is changed. The reason I
think that valid_tutor? isn't called is because I stuck "raise "ERROR"" into it
and that didn't come up in the "should require a valid tutor" spec.