[submenu] gallery2 -> gallery3
Summary:
We've finally got a working gallery3 setup on HKN with up-to-date photos.
Link to the new photo gallery instead of the old one.

Test Plan:
Click the link.

Revert Plan:
Go back to using gallery2 and revert this commit. If this happens bridge will be
extremely sad.

Tags:
gallery2, gallery3, submenu, link