Add officer challenges + hours to candidate portal
Challenges are handled as before.

Officer hours are handled as "Interactivity" event types, through
the regular event calendar. They are counted if they have rsvp's
confirmed - note that only VP has permissions to do so, which
means that unprivileged officers must message VP separatey to confirm RSVPs.

The total requirement of 3 is unchanged, but both challenges and hours
may count towards the total.